+++ isitfit/cli/cost.py
@@ -32,15 +32,17 @@ def analyze(ctx, filter_tags):
     from ..cost.utilizationListener import UtilizationListener
     from ..cost.cacheManager import RedisPandas as RedisPandasCacheManager
     from ..cost.datadogManager import DatadogCached
+    from ..cost.ec2TagFilter import Ec2TagFilter
     from isitfit.cost.ec2.reporter import ReporterAnalyzeEc2
 
     share_email = ctx.obj.get('share_email', None)
     ul = UtilizationListener(ctx)
     cache_man = RedisPandasCacheManager()
     ddg = DatadogCached(cache_man)
+    etf = Ec2TagFilter(filter_tags)
     ra = ReporterAnalyzeEc2()
     ra.set_analyzer(ul)
-    mm = MainManager(ctx, filter_tags, cache_man)
+    mm = MainManager(ctx, cache_man)
     def ra_postprocess_wrap(n_ec2_total, mm, n_ec2_analysed, region_include):
       ul.n_ec2_total = n_ec2_total
       ul.mm = mm
@@ -52,6 +54,8 @@ def analyze(ctx, filter_tags):
     ra_email_wrap = lambda *args, **kwargs: ra.email(share_email)
 
     # utilization listeners
+    mm.add_listener('ec2', etf.per_ec2)
+    mm.add_listener('ec2', mm._handle_ec2obj)
     mm.add_listener('ec2', ddg.per_ec2)
     mm.add_listener('ec2', ul.per_ec2)
     mm.add_listener('all', ul.after_all)
@@ -90,14 +94,16 @@ def optimize(ctx, n, filter_tags):
     from ..cost.optimizerListener import OptimizerListener
     from ..cost.cacheManager import RedisPandas as RedisPandasCacheManager
     from ..cost.datadogManager import DatadogCached
+    from ..cost.ec2TagFilter import Ec2TagFilter
     from isitfit.cost.ec2.reporter import ReporterOptimizeEc2
 
     ol = OptimizerListener(n)
     cache_man = RedisPandasCacheManager()
     ddg = DatadogCached(cache_man)
+    etf = Ec2TagFilter(filter_tags)
     ra = ReporterOptimizeEc2()
     ra.set_analyzer(ol)
-    mm = MainManager(ctx, filter_tags, cache_man)
+    mm = MainManager(ctx, cache_man)
 
     ra_display = lambda *args, **kwargs: ra.display()
     def ra_postprocess_wrap(n_ec2_total, mm, n_ec2_analysed, region_include):
@@ -109,6 +115,8 @@ def optimize(ctx, n, filter_tags):
 
     # utilization listeners
     mm.add_listener('pre', ol.handle_pre)
+    mm.add_listener('ec2', etf.per_ec2)
+    mm.add_listener('ec2', mm._handle_ec2obj)
     mm.add_listener('ec2', ddg.per_ec2)
     mm.add_listener('ec2', ol.per_ec2)
     mm.add_listener('all', ra_postprocess_wrap)