+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Version latest (0.15.0rc?, 2019-09-24?)
 - enh: split out report from ec2 optimizer
 - enh: share a context dict between `per_ec2` listeners instead of passing a list of objects
 - enh: move datadog out of mainManager and into listeners
+- enh: move `filter_tags` and `handle_ec2obj` out of mainManager ec2 loop and into listeners
 
 
 Version 0.14.0 (2019-11-15)